UBUNTU: SAUCE: vhost: make sure log_num < in_num
The code assumes log_num < in_num everywhere, and
that is true as long as in_num is incremented by
descriptor iov count, and log_num by 1.
However this breaks if there's a zero sized descriptor.

As a result, if a malicious guest creates a vring desc with desc.len = 0,
it may cause the host kernel to crash by overflowing
the log array. This bug can be triggered during the VM migration.

There's no need to log when desc.len = 0, so just don't increment
log_num in this case.

Fixes: 3a4d5c94e959 ("vhost_net: a kernel-level virtio server")
Reviewed-by: Lidong Chen <lidongchen@tencent.com>
Signed-off-by: ruippan <ruippan@tencent.com>
Signed-off-by: yongduan <yongduan@tencent.com>
Acked-by: Michael S. Tsirkin <mst@redhat.com>
Reviewed-by: Tyler Hicks <tyhicks@canonical.com>

CVE-2019-14835

(backported from email patch attachment)
[juergh: Adjusted context.]
Signed-off-by: Juerg Haefliger <juergh@canonical.com>