UBUNTU: SAUCE: s390/zcrypt: CEX7 toleration support
BugLink: https://bugs.launchpad.net/bugs/1848173

With the new z15 machine also came a new crypto card CEX7
which is backward compatible to CEX6 and CEX5. The newer
kernel code covers this with a QACT instruction which
automatically enables 'unknown' new crypto cards when
QACT tells that these are backward compatible.

However, the older kernels need at least some lines
of code mapping the CEX7 crypto hardware type back
to a known and compatible type ('toleration'). And
this is done with the patch here.

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>
Signed-off-by: Frank Heimes <frank.heimes@canonical.com>
Acked-by: Khalid Elmously <khalid.elmously@canonical.com>
Acked-by: Kleber Souza <kleber.souza@canonical.com>
Signed-off-by: Khalid Elmously <khalid.elmously@canonical.com>